chore: Split up Circle hooks
